refactor: DetailSectionPlace에서 pagination 로직을 PlaceCardList 컴포넌트로 분리
